+++ src/mruby_integration/audio/music.cpp
@@ -6,38 +6,6 @@
 #include "mruby_integration/struct_types.hpp"
 
 auto
-mrb_load_music_stream(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  char* path;
-  mrb_get_args(mrb, "z", &path);
-
-  auto* new_music = static_cast<Music*>(malloc(sizeof(Music)));
-  *new_music = LoadMusicStream(path);
-  mrb_value obj =
-    mrb_obj_value(Data_Wrap_Struct(mrb, Music_class, &Music_type, new_music));
-
-  setup_Music(mrb,
-              obj,
-              new_music,
-              new_music->ctxType,
-              new_music->looping,
-              new_music->frameCount);
-
-  return obj;
-}
-
-auto
-mrb_unload_music_stream(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  Music* music;
-  mrb_get_args(mrb, "d", &music, &Music_type);
-
-  UnloadMusicStream(*music);
-
-  return mrb_nil_value();
-}
-
-auto
 mrb_play_music_stream(mrb_state* mrb, mrb_value) -> mrb_value
 {
   Music* music;
@@ -146,16 +114,6 @@ mrb_get_music_time_played(mrb_state* mrb, mrb_value) -> mrb_value
 void
 append_audio_music(mrb_state* mrb)
 {
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "load_music_stream",
-                    mrb_load_music_stream,
-                    MRB_ARGS_REQ(1));
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "unload_music_stream",
-                    mrb_unload_music_stream,
-                    MRB_ARGS_REQ(1));
   mrb_define_method(mrb,
                     mrb->kernel_module,
                     "play_music_stream",